moving old tests to async + improving docstrings
